# Documents/api_definitions/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/api/public/windpower": {
            "get": {
                "summary": "Get Windpower",
                "description": "Get wind power production forecast.\nFetches forecast data from Fingrid dataset ID 245.\n\nReturns:\n    dict: A data point or an error message.",
                "operationId": "get_windpower_api_public_windpower_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FingridDataPoint"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/public/windpower/range": {
            "post": {
                "summary": "Post Windpower Range",
                "description": "Get wind power production data for a given time range.\nFetches data from Fingrid dataset ID 245.\n\nArgs:\n    time_range (TimeRangeRequest): Start and end time in RFC 3339 format.\n\nReturns:\n    list[dict]: A list of data points or an error message.",
                "operationId": "post_windpower_range_api_public_windpower_range_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TimeRangeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/FingridDataPoint"
                                    },
                                    "type": "array",
                                    "title": "Response Post Windpower Range Api Public Windpower Range Post"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/public/consumption": {
            "get": {
                "summary": "Get Consumption",
                "description": "Get electricity consumption forecast.\nFetches consumption data from Fingrid dataset ID 165.\n\nReturns:\n    dict: A data point or an error message.",
                "operationId": "get_consumption_api_public_consumption_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FingridDataPoint"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/public/consumption/range": {
            "post": {
                "summary": "Post Consumption Range",
                "description": "Get electricity consumption data for a given time range.\nFetches consumption data from Fingrid dataset ID 165.\n\nArgs:\n    time_range (TimeRangeRequest): Start and end time in RFC 3339 format.\n\nReturns:\n    list[dict]: A list of data points or an error message.",
                "operationId": "post_consumption_range_api_public_consumption_range_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TimeRangeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/FingridDataPoint"
                                    },
                                    "type": "array",
                                    "title": "Response Post Consumption Range Api Public Consumption Range Post"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/public/production": {
            "get": {
                "summary": "Get Production",
                "description": "Get electricity production forecast.\nFetches production data from Fingrid dataset ID 241.\n\nReturns:\n    dict: A data point or an error message.",
                "operationId": "get_production_api_public_production_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FingridDataPoint"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/public/production/range": {
            "post": {
                "summary": "Post Production Range",
                "description": "Get electricity production data for a given time range.\nFetches production data from Fingrid dataset ID 241.\n\nArgs:\n    time_range (TimeRangeRequest): Start and end time in RFC 3339 format.\n\nReturns:\n    list[dict]: A list of data points or an error message.",
                "operationId": "post_production_range_api_public_production_range_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TimeRangeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/FingridDataPoint"
                                    },
                                    "type": "array",
                                    "title": "Response Post Production Range Api Public Production Range Post"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/public/weather": {
            "post": {
                "summary": "Post Weather",
                "description": "Get weather forecast for a specific UTC time and location (POST version).\nFetches weather forecast data from the MET API.\n\nArgs:\n    request (WeatherRequest): Contains latitude, longitude, and timestamp in RFC 3339 format.\n\nReturns:\n    dict: Weather forecast data or an error message.",
                "operationId": "post_weather_api_public_weather_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WeatherRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WeatherDataPoint"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/public/price/range": {
            "post": {
                "summary": "Post Price Range",
                "description": "Get price data for specific time range from Porssisahko API\n\nArgs:\n    Timestamp in RFC 3339 format.\n\nReturns:\n    list[dict]: Price data or an error message",
                "operationId": "post_price_range_api_public_price_range_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TimeRangeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/public/data": {
            "get": {
                "summary": "Get Prices",
                "operationId": "get_prices_api_public_data_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/PriceDataPoint"
                                    },
                                    "type": "array",
                                    "title": "Response Get Prices Api Public Data Get"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/public/data/today": {
            "get": {
                "summary": "Get Prices Today",
                "operationId": "get_prices_today_api_public_data_today_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/PriceDataPoint"
                                    },
                                    "type": "array",
                                    "title": "Response Get Prices Today Api Public Data Today Get"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "summary": "Register",
                "operationId": "register_api_auth_register_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "summary": "Login",
                "operationId": "login_api_auth_login_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "summary": "Logout",
                "operationId": "logout_api_auth_logout_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string",
                        "title": "Error",
                        "description": "Error message describing the issue.",
                        "example": "An error occurred"
                    }
                },
                "type": "object",
                "required": [
                    "error"
                ],
                "title": "ErrorResponse"
            },
            "FingridDataPoint": {
                "properties": {
                    "startTime": {
                        "type": "string",
                        "title": "Starttime",
                        "description": "Start time in RFC 3339 format (e.g., 2024-05-01T00:00:00Z)",
                        "example": "2024-05-01T00:00:00Z"
                    },
                    "endTime": {
                        "type": "string",
                        "title": "Endtime",
                        "description": "End time in RFC 3339 format (e.g., 2024-05-02T00:00:00Z)",
                        "example": "2024-05-02T00:00:00Z"
                    },
                    "value": {
                        "type": "number",
                        "title": "Value",
                        "description": "Value of the data point",
                        "example": "7883.61"
                    }
                },
                "type": "object",
                "required": [
                    "startTime",
                    "endTime",
                    "value"
                ],
                "title": "FingridDataPoint"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "PriceDataPoint": {
                "properties": {
                    "startDate": {
                        "type": "string",
                        "title": "Startdate",
                        "description": "UTC str in RFC 3339 format",
                        "example": "2025-05-08T04:00:00.000Z"
                    },
                    "price": {
                        "type": "number",
                        "title": "Price",
                        "description": "Floating-point number representing the price in euro cents",
                        "example": 0.61
                    }
                },
                "type": "object",
                "required": [
                    "startDate",
                    "price"
                ],
                "title": "PriceDataPoint"
            },
            "TimeRangeRequest": {
                "properties": {
                    "startTime": {
                        "type": "string",
                        "title": "Starttime",
                        "description": "Start time in RFC 3339 format (e.g., 2024-05-01T00:00:00Z)",
                        "example": "2024-05-01T00:00:00Z"
                    },
                    "endTime": {
                        "type": "string",
                        "title": "Endtime",
                        "description": "End time in RFC 3339 format (e.g., 2024-05-02T00:00:00Z)",
                        "example": "2024-05-02T00:00:00Z"
                    }
                },
                "type": "object",
                "required": [
                    "startTime",
                    "endTime"
                ],
                "title": "TimeRangeRequest"
            },
            "User": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    },
                    "role": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Role",
                        "default": "user"
                    }
                },
                "type": "object",
                "required": [
                    "email",
                    "password"
                ],
                "title": "User"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "WeatherDataPoint": {
                "properties": {
                    "temperature_celsius": {
                        "type": "number",
                        "title": "Temperature Celsius",
                        "description": "Air temperature in degrees Celsius based on the forecast.",
                        "example": 8.5
                    },
                    "wind_speed_mps": {
                        "type": "number",
                        "title": "Wind Speed Mps",
                        "description": "Wind speed in meters per second based on the forecast.",
                        "example": 2.2
                    },
                    "closest_forecast_time": {
                        "type": "string",
                        "title": "Closest Forecast Time",
                        "description": "Timestamp (in UTC, RFC 3339 format) of the closest available weather forecast.",
                        "example": "2025-05-09T16:00:00Z"
                    }
                },
                "type": "object",
                "required": [
                    "temperature_celsius",
                    "wind_speed_mps",
                    "closest_forecast_time"
                ],
                "title": "WeatherDataPoint"
            },
            "WeatherRequest": {
                "properties": {
                    "lat": {
                        "type": "number",
                        "title": "Lat",
                        "description": "Latitude",
                        "example": 60.1699
                    },
                    "lon": {
                        "type": "number",
                        "title": "Lon",
                        "description": "Longitude",
                        "example": 24.9384
                    },
                    "timestamp": {
                        "type": "string",
                        "title": "Timestamp",
                        "description": "UTC str in RFC 3339 format, e.g. 2024-05-05T13:30:00Z",
                        "example": "2024-05-05T13:30:00Z"
                    }
                },
                "type": "object",
                "required": [
                    "lat",
                    "lon",
                    "timestamp"
                ],
                "title": "WeatherRequest"
            }
        }
    }
}